Clean up command autocomplete
I thought it'd be better to abstract this functionality to a few
functions in scriptfunc, for if a plugin wants to hook the list of
commands cleanly (ex. fakecommands) or would like to get a list of
autocomplete commands without copy-pasting the code (which would work
well with, again, plugins hooking the list of possible commands).