Prevent adding .vvvvvv files from assets to ZIPs
This was a little oversight: if you had a .vvvvvv file in your level
assets folder, and then made a ZIP using Ved, it would add both
level files to the ZIP, no matter if that meant there were now two
files with the same name in one folder. Oopsie...

Now, all .vvvvvv files from within the assets folder will not be
included. (Even if the names are different it would still probably not
be intentional and cause confusion)

[pre12]