Replace usages of \u escape by \x
I needed a U+200E in a couple places, so I used a Unicode escape, which
worked. Until I tried to boot Ved in a somewhat older LÖVE where it was
a syntax error. This is apparently build-dependent or something because
it worked fine on my Linux LÖVE 0.9.1, but oh well. Raw byte escapes
are compatible with everything.

[pre13]