Allow placing terminals with empty script name
This is used for decorative terminals without an activity zone. You
used to have to go into the entity properties to remove the name, now
you can simply press enter.

(It also allows script boxes and roomtext without text, but we'll
assume the user knows what they want and it's not the end of the world
if there's a pointless entity somewhere.)

[pre17]