Redesign flags screen for bigger fonts
Flag names should be displayed in the level font, and as such, they
have to be able to be displayed in fonts higher than 8 pixels high.
Which meant I had to change the current two-column layout into a
three-column layout.

It feels like this needs a little more tweaking - a four-column layout
would feel more satisfying I think, being 25-25-25-25 instead of
34-34-32, but that gets us into trouble with string lengths so it'll do
for now.

Also did a bit of migration away from ved_print[f] functions - we need
to get rid of all of them!

[pre18]