Stop flag drag-and-drop when reentering flags list
If you click and drag a flag in the flags list, and then decide you
don't want to move any flags, so you press Esc, and then re-enter the
flags list, the flag drop would still be in progress and actually drop
the flag wherever your mouse was. This is now fixed.

[pre18]