Allow pressing Esc to cancel placing script box
If you click to place one corner of a script box, it should be possible
to cancel by pressing Esc. This works now. You can also switch tools or
go to a different room to cancel placing a script box.

Also, it was actually pretty easy to break things a little by pressing
buttons while Ved is waiting for you to press the second corner -
switching between the hidden script box subtools wasn't properly
blocked, you could place the second corner in a different room and get
weird script box dimensions (as well as an invisible name prompt
because it was in a different room), and some other things. This should
work a lot better now.

[pre18]