Add clicking on entities to take certain actions
This is a feature Ved never really needed since it had subtools to
immediately place an entity facing the "correct" direction, and
otherwise you could always use the button in the right click menu to
change it. But it might actually be more comfortable to do it the main
editor way - so you can now also click entities to do certain things
with them in Ved, just like you can in the main editor!

For example, clicking a gravity line's control point to change it
between horizontal and vertical, clicking rescuable crewmates to change
their color, and so forth. Also, something the main editor doesn't
have: clicking a warp token to immediately go to the other end!

The only thing I didn't do was being able to flip terminals this way -
it already prompts to change the script name, and doing both at the
same time like the main game does feels wrong. I think, ideally, you'd
want a click to edit the script name, and then while editing the script
name, you can freely click to flip the terminal however many times you
want (and then just press Esc or Enter to cancel editing the name). But
that is more complicated to implement, so for now you can always use
the right click menu option.

[pre19]