Fix cursor stuck as I-beam after closing input
If your cursor enters a text field, it becomes an I, and if your cursor
leaves the text field, it goes back to the pointer. However, if you
press Enter/Esc to stop editing roomtext, a script name or a roomname
while your cursor was on the editable text, it would not revert to the
pointer, and be stuck on the I-beam for a little while until something
reset it, like going to a different state or opening another input.
This is now fixed.

[pre21]