Add quick Reload button to graphics viewer
This allows you to externally edit a graphics file, and then instantly
reload without using the files picker, and without losing your view
options like position and zoom.

The shortcut is R, which takes the shortcut previously used by the RGB
checkbox, so that is now C (for Color). Sorry if that was in your
muscle memory! But R seems like the only intuitive shortcut I could use
for Reload, when L is already taken for Load.

(I might've accidentally moved up these non-Elements interface elements
twice as many pixels as I wanted because I was taking dimensions from a
screenshot of 2x-scaled Ved... oops! But it feels like a happy accident
because I think it probably looks better this way, lol)

[pre23]