Migrate all uses of deprecated ved_print*/font8
They should now all be using the font objects reserved for each
specific purpose: font_ui for text that's part of the interface and
needs to be able to display the current Ved language, font_level for
text that's from the level and needs to be displayed in the level font,
and font_8x8 for text that should always be displayed in the 8x8 font.
As well as possible special cases of indexing fonts_main directly.

This has also fixed or alleviated several bugs in the process, where
things were being printed using the wrong font. Though, there are still
some issues to solve, like the level compare screen only being in one
font even though multiple may be needed, and the metadata preview on
the loading screen now uses the correct fonts but lines may overlap a
bit with CJK levels.

[pre24]