Fix plugin .zip files with dots in name breaking
Hypothetically, if you have a plugin like green_mode_1.1.zip, this
would break when trying to load its UI files because the dot in 1.1
conflicts with the module name separator (err, whatever it's called)
for `require`.

I feel like this is something that broke with the semi-recent change
of all our `require`s to use dots instead of slashes? Since I feel like
I've been sharing plugin zips with version numbers several times before
without issues. But it's probably that the issue used to only present
itself if a plugin used includes (which is very rare) since plugins
having UIs didn't exist yet for the longest time, and the plugin which
unexpectedly broke when I shared it had a UI.

This affects both zip files and folders by the way, but fixing it for
zip files was easy since we can just replace dots by underscores in the
name when it is mounted. Fixing it for folders sounds like a pretty
tall order, but the problem should also be much less common there since
zip files are for packaged versions of plugins (hence you might want to
use a version number in the filename) while folders would commonly be
for (unversioned) development versions of plugins.

[pre02]