Fix "#" being added for blank lines in L textboxes
Blank lines in internal scripting mode are replaced with # symbols to
prevent issues. In textbox contents, blank lines should stay blank
instead. This part wasn't working in textboxes where the number of
lines has an L to allow the number to be higher than 11 - e.g.
text(gray,0,0,15L). This is now fixed.

[pre01]