Always convert flag names inside textbox contents
The script compilation function (run when leaving the script editor)
accounted for which lines were a part of textboxes (and as a result
didn't process flag commands that were part of textboxes), while the
decompilation function (run when opening the script) didn't. This
resulted in Ved converting flag numbers to their names when opening the
script, but not converting them back to numbers when saving, as
reported by stupid cat. This was a problem for specific cases where
textbox contents are actually run by the game, because the game doesn't
expect flag names which are now forced to be in the raw script.

My solution is to make the script compiler convert flag names to
numbers even in textbox contents. This makes it consistent with
script decompilation and swapping flags, which both didn't keep track
of this either.

You'd think ideally those functions would be made "smarter" too - where
script decompilation only converts flag numbers to names outside of
textbox contents, in case someone were to actually want to use the
exact syntax of a flag command inside text box contents displayed to
the player. But then you would never be able to use flag names in what
Ved thinks are textbox contents, and the code would need to be more
complex for not much practical gain.

[pre02]