Change LMB on terminal/script box to edit script
Left clicking on a terminal or script box entity used to mimic the main
editor by letting you modify the script name.

I'm trying out another action: left clicking on the terminal/script box
will now open the script in the script editor. When working on my level
I find myself having to constantly right click > Edit script to extend
the cutscene I was working on, or to check how I did something in one
place so I can do a similar thing somewhere else, whereas changing a
script name (detaching it from the script contents it's set to) is a
more rare occurrence that I would *expect* to have to right click for.
If you were uncertain what to name a script, it's better to rename it
from the script list anyway, to mass-rename all references.

[pre03]