Change flags list to a 4-column layout
Flashback to c47373a01a2135c5600e74b0aa65a157f4533a8e: I had to change
the layout of the flags list from a two-column layout (0-49, 50-99) to
a three-column layout (0-33, 34-67, 68-99). This layout is a little
awkward and doesn't lend well to organizing flags around round
numbers, so I'm now making it a four-column layout instead
(0-24, 25-49, 50-74, 75-99) with less wide but higher buttons. I had to
sacrifice the "Used"/"Not used" label to make it fit, but that's also
shown by the button colors, so it was kinda redundant anyway. And the
screen looks cleaner with it removed.

I also added the total number of used flags in the top right corner so
it might be a little bit clearer what the backgrounds mean now to
anyone unaware, and because it might be a useful number to have.

[pre05]