Update file size display function
This is the function that shows how many bytes/kB/MB/etc something is
(currently only used in the music player/editor, but I'm planning to
use it in the file picker too).

Visibly, this change is pretty small - it used to always give one
decimal place (e.g. 3.2 MB and 333.2 MB), now it always uses 3
digits (3.22 MB, 33.2 MB, 333 MB) for better consistency. It can now
also display terabytes (mostly because I may add it to the drive list),
and numbers of bytes are now listed as B instead of a byte/bytes plural
string (mostly so I need to worry less about the width of the column
but it also looks a bit better in a list with e.g. both B and kB
I think). Also, the decimal point will be changeable per language.

[pre01]